ai-tutor-api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pyjwt>=2.9.0
python-dotenv>=1.0.1
httpx>=0.28.0
pydantic-settings>=2.7.0
sse-starlette>=2.2.0
openai>=1.60.0
langsmith>=0.1.0
supabase>=2.11.0
tiktoken>=0.7.0
pymupdf>=1.24.0
python-docx>=1.1.0
tenacity>=9.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
